tidy: fix Kotlin compilation warnings
Replace deprecated -Xjvm-default=all string arg with typed JvmDefaultMode.NO_COMPATIBILITY property.
Narrow java.lang.Double star import to the two static methods actually used.
Remove unnecessary safe calls on non-null receivers in MemoryCache and LogProcessor.
Remove redundant .toInt() conversion in NumericBin where Short arithmetic already yields Int.
Replace deprecated URL(String) constructor with URI(String).toURL() in YamlSerdeTest.
Fix processRecords parameter name to match Subscriber supertype in LocalLogTest.
Rename test methods containing % to use 'percent-' for Windows filesystem compatibility.

+++ core/src/main/kotlin/xtdb/pg/codec/NumericBin.kt
@@ -61,7 +61,7 @@ fun decode(bb: ByteBuffer): BigDecimal {
     var num = BigInteger.ZERO
 
     for (i in 0 until digitsNum) {
-        val base = TEN_THOUSAND.pow((digitsNum - i - 1).toInt())
+        val base = TEN_THOUSAND.pow(digitsNum - i - 1)
         val digit = BigInteger.valueOf(bb.short.toLong())
         num = num.add(base.multiply(digit))
     }
